probing media and generating thumbnails is now handled by ffmpeg

+++ VideoLib/DashVideoDecoder.h
@@ -207,7 +207,7 @@ namespace VideoLib {
 			} 
 		}
 
-		int getSizeBytes() const {
+		int64_t getSizeBytes() const {
 
 			return(videoDecoder->getSizeBytes() + audioDecoder->getSizeBytes());
 		}
